# Dipolarkernal speedup