--- a/docsrc/source/changelog.rst
+++ b/docsrc/source/changelog.rst
@@ -27,6 +27,7 @@ Release Notes
 Release ``v1.1.3`` - Ongoing
 ------------------------------------------
 - |fix| : Removes unnecessary files from the docs
+- |efficiency| : Improves the performance of the ``dipolarkernel`` function by 10-30% (:pr:`473`), by caching the interpolation of he effective dipolar evolution time vector.
 
 Release ``v1.1.2`` - November 2023
 ------------------------------------------